DBZ-7024 Move creation of converters into dedicated class
